requirements.txt
autogen-agentchat
autogen-core  
autogen-ext
openai
tiktoken
python-dotenv
requests
fastapi
uvicorn[standard]
pydantic
python-multipart
jinja2